Rename files to PascalCase for consistency with other files
