+++ lib/DependencyProviderOffline.js
