+++ lib/Solve.js
@@ -6,7 +6,7 @@ const path = require('path');
 const ElmJson = require('./ElmJson');
 const Project = require('./Project');
 
-let depsProvider = require('./dependency-provider.js');
+let depsProvider = require('./DependencyProvider.js');
 let wasm = require('elm-solve-deps-wasm');
 wasm.init();
